Pdf-to-PPTX/requirements.txt
Flask>=3.0.0,<4.0.0
gunicorn>=21.2.0,<22.0.0
pdf2image>=1.17.0,<2.0.0
python-pptx>=0.6.23,<0.7.0
Pillow>=10.0.0,<12.0.0
